GAME: Replace hardcoded Dead Player Body spritelist with `SPR2_` sprite set functionality
The Dead Player Body sprites are no longer hardcoded, meaning Custom Characters can also have a Dead Body sprite now. The game will try to load a sprite lump named as "MDEDA0" ("SPR2_MDED" as Sprite2) and place it on the Dead Player Body object. If the sprite is not found the body defaults to the Skeleton

Bob and Superguide sprites were also removed