MISC: Introduce a prototype of the new MM setup utility/panel (`MMSETUP` CCMD)
The new MM setup panel is meant to make the configuration process of the MM game easier in some way or another. In a nutshell, it is a UI for changing CVARs that you always could change yourself with a console.

The design of the menu might be changed and, perhaps, CVARs themselves would be removed too, if there wouldn't be a need for them