MAPS: Add a new variant of the `MAPKB` ("Railway Station")
Update made by @mm8573 that turns the map from "Chill & colorful Railway Station" to "1984-like station controlled by Eggman". Because it is a drastic change in map's atmosphere that can turn out to be a bad design change, the files of the old map variant are kept for an easy reverting process.

Basically we cannot decide what to do with this change.

There is a concept of implementing the Time Travel mechanic, much like "East City" (`MAPKD`), to merge both map designs