CODE & HUD: Remove UTF-8 support for chat and improve readability in 320x200 video mode
Tests have shown that SRB2's LUA is extremely limited for things like UTF-8 compatible chat, even my workarounds fail. There can be an UTF-8 chat support added with LUA, but it is very limited and is just not worth it.

Initially this commit was going to add Unifont into the game, but since UTF-8 chat is removed, there is no need for it. Some readability features meant for Unicode are now used instead for the vanilla font